# api/cmd/api/openapi.json
{
  "components": {
    "schemas": {
      "CreateSandboxRequest": {
        "description": "Sandbox create request. CPU/memory may be overridden by the template's baked snapshot dimensions.",
        "properties": {
          "cpu": {
            "description": "Requested vCPUs",
            "type": "integer"
          },
          "disk_gb": {
            "description": "Optional rootfs size in GiB",
            "type": "integer"
          },
          "from_snapshot": {
            "description": "Restore from this snapshot ID instead of booting the template",
            "type": "string"
          },
          "memory_mb": {
            "description": "Requested memory in MiB",
            "type": "integer"
          },
          "metadata": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Free-form string metadata; keys starting with secret. are write-only",
            "type": "object"
          },
          "persistent": {
            "description": "true exempts the sandbox from the idle reaper",
            "type": "boolean"
          },
          "template": {
            "description": "Template name (e.g. base, code-interpreter, agent, browser). Public templates or ones owned by the caller's workspace only",
            "type": "string"
          },
          "ttl_seconds": {
            "description": "Idle TTL in seconds; null uses the platform default",
            "type": [
              "integer",
              "null"
            ]
          },
          "volumes": {
            "description": "Durable volumes to attach",
            "items": {
              "$ref": "#/components/schemas/VolumeMount"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "description": "Human-readable error message",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "SuspendedError": {
        "properties": {
          "detail": {
            "description": "Human-readable explanation",
            "type": "string"
          },
          "error": {
            "description": "Short error code message",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "pds_... API token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "url": "https://pandastack.ai"
    },
    "description": "REST API for PandaStack — Firecracker microVM sandboxes, managed Postgres, git-driven apps, serverless functions, and outbound webhooks. Authenticate with a pds_ bearer token from the dashboard's API Tokens page.",
    "title": "PandaStack API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/sandboxes": {
      "get": {
        "operationId": "listSandboxes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Sandbox"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Array of sandboxes owned by the caller's workspace"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "List failed"
          }
        },
        "summary": "List the caller's sandboxes (workspace-scoped; internal pool/tombstone rows filtered out)",
        "tags": [
          "Sandboxes"
        ]
      },
      "post": {
        "operationId": "createSandbox",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSandboxRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sandbox"
                }
              }
            },
            "description": "Sandbox created and running"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Malformed JSON body or invalid template name"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuspendedError"
                }
              }
            },
            "description": "Account suspended, or template owned by another workspace"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Create failed"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "No available compute node"
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Host memory admission refused (control plane places elsewhere)"
          }
        },
        "summary": "Create a sandbox (snapshot-restore boot, ~180ms p50)",
        "tags": [
          "Sandboxes"
        ]
      }
    },
...
    "/v1/templates": {
      "get": {
        "operationId": "listTemplates",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Template"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Bare JSON array of templates (globals sorted first, then by name)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Catalog query failed"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List the template catalog (global first-party templates plus the caller's workspace-owned custom templates)",
        "tags": [
          "Templates"
        ]
      }
    },
    "/v1/templates/build": {
      "post": {
        "operationId": "buildTemplate",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TemplateBuildRequest"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/TemplateBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateBuild"
                }
              }
            },
            "description": "Content-hash dedup hit: an identical successful build already exists for this workspace; its terminal record is returned without rebuilding"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateBuild"
                }
              }
            },
            "description": "Build accepted and running asynchronously. Server-side (dockerfile) builds return an id prefixed \"tb_\"; agent-direct builds return the agent build state"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Invalid name ([a-zA-Z0-9._-], max 64), size_mb > 16384, cpu outside 1..64, memory_mb outside 128..65536, unreadable build-context file, or neither image_ref nor rootfs provided"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Caller has no workspace resolved"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Replace refused: template is public/seeded, or owned by another workspace"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Template already exists and replace=true was not passed"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Build could not be recorded or staged"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "dockerfile path requested but server-side build is not enabled on this deployment"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Build a custom template. Three mutually-exclusive inputs: dockerfile (server-side Cloud Build, self-serve), image_ref (agent pulls an OCI image from the trusted registry), or rootfs (legacy inline tarball upload). Async: poll GET /v1/templates/builds/{id}",
        "tags": [
          "Templates"
        ]
      }
    },
    "/v1/templates/builds": {
      "get": {
        "operationId": "listTemplateBuilds",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TemplateBuild"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Bare JSON array of build states"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List agent-side template builds (proxied to an agent; in-memory build ledger, unordered)",
        "tags": [
          "Templates"
        ]
      }
    },
    "/v1/templates/builds/{id}": {
      "get": {
        "operationId": "getTemplateBuild",
        "parameters": [
          {
            "description": "Build id (\"tb_…\" for server-side builds, agent-generated id otherwise)",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateBuild"
                }
              }
            },
            "description": "Build status record"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Build not found (or owned by another workspace)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get one build's status. Ids prefixed \"tb_\" (server-side Cloud Build) are served from the control-plane DB (workspace-scoped); other ids proxy to the agent's in-memory ledger",
        "tags": [
          "Templates"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://api.pandastack.ai"
    }
  ]
}